+++ src/Ribbon/index.d.ts
@@ -1,0 +1,4 @@
+import type { SvelteComponentTyped } from 'svelte'
+export class Ribbon extends SvelteComponentTyped<{
+	corner?:boolean
+}> {}